extract map and sidebar components
